be/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
SQLAlchemy==2.0.19
pymysql==1.1.0
python-dotenv==1.0.0
httpx==0.24.1
pydantic==1.10.9
passlib[bcrypt]==1.7.4
rapidfuzz==2.13.7
python-jose[cryptography]==3.3.0
